INS-593: move scenario to separate file
